Feat : 상세보기(새창으로 띄우기) #2

+++ admin/src/main/java/com/sh/admin/order/model/dto/OrderDto.java
@@ -17,7 +17,7 @@ public class OrderDto {
     private LocalDate orderDate; // 주문날짜
     private int sumPrice; // 주문합계
     private OrderStatus orderStatus; // 주문상태 (주문요청, 발송완료, 재고부족, 주문취소, 환불완료)
-    private int trackNumber; // 송장번호
+    private long trackNumber; // 송장번호
     private LocalDate shipmentDate; // 발송날짜
 
     private List<OrderItemDto> orderItemDtos;
